TTT-10: Reorganises source code folders to impose module namespacing. Separates tests from main line code.
